% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtcars2.R
\docType{data}
\name{mtcars2}
\alias{mtcars2}
\title{mtcars2: Motor Trend Car Road Tests}
\format{
A data frame with 32 observations on 11 (numeric) variables.
\describe{
\item{mpg}{Miles/(US) gallon}
\item{cyl}{Number of cylinders}
\item{disp}{Displacement (cu.in.)}
\item{hp}{Gross horsepower}
\item{drat}{Rear axle ratio}
\item{wt}{Weight (1000 lbs)}
\item{qsec}{1/4 mile time}
\item{vs}{Engine (0 = V-shaped, 1 = straight)}
\item{am}{	Transmission (0 = automatic, 1 = manual)}
\item{gear}{Number of forward gears}
\item{carb}{Number of carburetors}
}
}
\source{
mtcars function in datasets package.
}
\usage{
mtcars2
}
\description{
Cópia do mtcars para demonstracao.
}
\keyword{datasets}
