% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/substitui_nome.R
\name{substitui_nome}
\alias{substitui_nome}
\title{substitui_nome}
\usage{
substitui_nome(
  string,
  aluno = "Victor",
  aluna = "Samantha",
  professor = "Rafael"
)
}
\arguments{
\item{string}{Uma string}

\item{aluno}{Um nome de aluno, padrao "Victor"}

\item{aluna}{Um nome de aluna, padrao "Samantha"}

\item{professor}{Um nome de professor, padrao "Rafael"}
}
\value{
A string com palavras "professor","aluno" ou "aluna" substituidas
por nomes "Rafael", "Victor" ou "Samantha" respectivamente.
}
\description{
Pacote teste para a disciplina de Perspectiva em Ciência de Dados que substitui nomes dentro de uma string.
}
\examples{
substitui_nome("O professor deu uma prova. A aluna foi muito bem, ja o aluno, nem tanto.")

}
